Read-only legarto application framework system tree
Change the /legato bind to /mnt/legato
Update the startlegato.sh to work with standalone read-only
legato system tree
Update legato-image reciepe to handle the legato read-only
system tree

Resolves: LXSWI9X40-24
Related: LE9XXX-242

+++ common/recipes-legato/legato-af/legato-af_git.bb
@@ -96,7 +96,11 @@ do_install_image() {
     # legato-image
     for target in ${LEGATO_ROOTFS_TARGETS}; do
         mkdir -p ${LEGATO_STAGING_DIR}/$LEGATO_VERSION/${target}
-        cp -R ${B}/build/${target}/staging/* ${LEGATO_STAGING_DIR}/$LEGATO_VERSION/${target}/
+        if [ -d ${B}/build/${target}/readOnlyStaging/legato ]; then
+            cp -R ${B}/build/${target}/readOnlyStaging/legato/* ${LEGATO_STAGING_DIR}/$LEGATO_VERSION/${target}/
+        else
+            cp -R ${B}/build/${target}/staging/* ${LEGATO_STAGING_DIR}/$LEGATO_VERSION/${target}/
+        fi
     done
 }
